moved events page components to their own subfolder and updated relevant import statements
